Move mock generate commands to non-integration-tagged files
